INTYGFV-15731: Updated name of GetActiveSickLeavesService to GetActiveSickLeavesResponseService and GetSickLeavesService.java to GetActiveSickLeavesService.
